+++ CHANGELOG.md
@@ -6,8 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+* cleanup backends
+
 ### Added
 * a `user_id` column into a `Logs` table
+* a `ipaddr` column into a `Logs` table
 * tests